Wrap UnifiedFooter to fix locale options workaround
